# Optimized destination position calculation in invert_entity()