### Brief description of the changes:

1. Skipped `src_pos.x == dest_pos.x` check, as it is always true
2. Replaced `src_pos.y == dest_pos.y` with `dest_pos.y == 0`, as it can be true only in this case
3. Removed unused `src_pos`
4. Reordered so that `return` check is made as soon as possible
5. Replaced `local source = event.source` declaration used only once in `local dest_pos = source.position`, with direct `local dest_pos = event.source.position`

### Benchmark
3600 tick profile right after map restart:
```
before:
117502x "invert_entity". Total Duration: 19746.395700ms avg: 0.16805157ms
	1527712x "__index". Total Duration: 7555.191300ms
	117502x "teleport". Total Duration: 912.231100ms
	582x "__newindex". Total Duration: 9.324700ms
	106x "table_insert". Total Duration: 0.573600ms
after:
129379x "invert_entity". Total Duration: 20141.495500ms avg: 0.15567824ms
	1552671x "__index". Total Duration: 7659.959300ms
	129379x "teleport". Total Duration: 812.593500ms
	428x "__newindex". Total Duration: 6.881100ms
	70x "table_insert". Total Duration: 0.358600ms
```
Around 7% improvement

### Tested Changes:
- [x] I've tested the changes locally.
- [ ] I've not tested the changes.